Refactor to remove package tangle.
