+++ project-set/components/client-auth/src/main/java/com/rackspace/papi/components/clientauth/ClientAuthenticationHandlerFactory.java
@@ -2,6 +2,7 @@ package com.rackspace.papi.components.clientauth;
 
 import com.rackspace.papi.commons.config.manager.UpdateListener;
 import com.rackspace.papi.commons.util.regex.KeyedRegexExtractor;
+import com.rackspace.papi.components.clientauth.common.AuthenticationHandler;
 import com.rackspace.papi.components.clientauth.common.UriMatcher;
 import com.rackspace.papi.components.clientauth.config.ClientAuthConfig;
 import com.rackspace.papi.components.clientauth.config.URIPattern;
