+++ project-set/components/client-auth/src/main/java/com/rackspace/papi/components/clientauth/common/AuthenticationHandler.java
@@ -1,4 +1,4 @@
-package com.rackspace.papi.components.clientauth;
+package com.rackspace.papi.components.clientauth.common;
 
 import com.rackspace.auth.AuthGroup;
 import com.rackspace.auth.AuthToken;
@@ -8,9 +8,6 @@ import com.rackspace.papi.commons.util.http.HttpStatusCode;
 import com.rackspace.papi.commons.util.regex.ExtractorResult;
 import com.rackspace.papi.commons.util.regex.KeyedRegexExtractor;
 import com.rackspace.papi.commons.util.servlet.http.ReadableHttpServletResponse;
-import com.rackspace.papi.components.clientauth.common.AuthTokenCache;
-import com.rackspace.papi.components.clientauth.common.Configurables;
-import com.rackspace.papi.components.clientauth.common.UriMatcher;
 import com.rackspace.papi.filter.logic.FilterAction;
 import com.rackspace.papi.filter.logic.FilterDirector;
 import com.rackspace.papi.filter.logic.common.AbstractFilterLogicHandler;
