+++ project-set/components/client-auth/src/main/java/com/rackspace/papi/components/clientauth/rackspace/v1_1/RackspaceAuthenticationHandler.java
@@ -5,7 +5,7 @@ import com.rackspace.auth.AuthToken;
 import com.rackspace.auth.rackspace.AuthenticationService;
 import com.rackspace.papi.commons.util.regex.ExtractorResult;
 import com.rackspace.papi.commons.util.servlet.http.ReadableHttpServletResponse;
-import com.rackspace.papi.components.clientauth.AuthenticationHandler;
+import com.rackspace.papi.components.clientauth.common.AuthenticationHandler;
 import com.rackspace.papi.components.clientauth.common.AuthTokenCache;
 import com.rackspace.papi.components.clientauth.common.Configurables;
 import com.rackspace.papi.components.clientauth.common.UriMatcher;
