+++ project-set/components/client-auth/src/main/java/com/rackspace/papi/components/clientauth/rackspace/v1_1/RackspaceAuthenticationHandlerFactory.java
@@ -3,7 +3,7 @@ package com.rackspace.papi.components.clientauth.rackspace.v1_1;
 import com.rackspace.auth.rackspace.AuthenticationService;
 import com.rackspace.auth.rackspace.AuthenticationServiceFactory;
 import com.rackspace.papi.commons.util.regex.KeyedRegexExtractor;
-import com.rackspace.papi.components.clientauth.AuthenticationHandler;
+import com.rackspace.papi.components.clientauth.common.AuthenticationHandler;
 import com.rackspace.papi.components.clientauth.common.AuthTokenCache;
 import com.rackspace.papi.components.clientauth.common.Configurables;
 import com.rackspace.papi.components.clientauth.common.UriMatcher;
